feat: add bytecode replacement based on function selector
- Implement precompile resolution using first 4 bytes of input
- Move multicall to contracts/precompiles as first precompile
