+++ crates/contracts/Makefile
@@ -1,5 +1,5 @@
-PRECOMPILES=evm blake2 sha256 ripemd160 identity modexp ecrecover
-EXAMPLES=greeting fairblock multicall
+PRECOMPILES=evm blake2 sha256 ripemd160 identity modexp ecrecover multicall
+EXAMPLES=greeting fairblock
 CUR_DIR="$(shell pwd)"
 OUT_DIR="${CUR_DIR}/assets"
 RUSTFLAGS='-C link-arg=-zstack-size=131072 -C target-feature=+bulk-memory'
