+++ crates/contracts/src/lib.rs
@@ -11,6 +11,10 @@ mod evm;
     feature = "ripemd160",
     feature = "identity",
     feature = "modexp",
-    feature = "ecrecover",
+    feature = "ecrecover"
 ))]
 mod precompile;
+
+/// Native smart contracts
+#[cfg(any(feature = "multicall"))]
+mod precompiles;
