+++ crates/contracts/src/precompiles/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "multicall")]
+pub mod multicall;
