+++ crates/core/src/blended.rs
@@ -1,5 +1,4 @@
 mod evm;
-mod multicall;
 mod syscall;
 mod util;
 mod wasm;
@@ -8,7 +7,6 @@ use crate::{
     types::NextAction,
 };
 use alloc::boxed::Box;
-use fluentbase_genesis::EXAMPLE_MULTICALL_ADDRESS;
 use fluentbase_sdk::{
     bytes::BytesMut,
     codec::FluentABI,
@@ -25,9 +23,9 @@ use fluentbase_sdk::{
     SovereignAPI,
     SovereignContextReader,
     SyscallInvocationParams,
+    PRECOMPILE_MULTICALL,
     STATE_MAIN,
 };
-use multicall::MULTICALL_SELECTOR;
 use revm_interpreter::{
     CallInputs,
     CallOutcome,
@@ -96,17 +94,12 @@ impl<SDK: SovereignAPI> BlendedRuntime<SDK> {
         contract_context: &ContractContext,
         params: SyscallInvocationParams,
     ) -> NextAction {
-        // Check if this is a multicall invocation
-        let is_multicall = params.input.len() >= 4 && params.input[..4] == MULTICALL_SELECTOR;
-
-        // Prepare shared context input
         let context_input = SharedContextInputV1 {
             block: self.sdk.context().clone_block_context(),
             tx: self.sdk.context().clone_tx_context(),
             contract: contract_context.clone(),
         };
 
-        // Encode context input with params
         let mut buf = BytesMut::new();
         FluentABI::encode(&context_input, &mut buf, 0).unwrap();
         buf.extend_from_slice(params.input.as_ref());
@@ -119,41 +112,20 @@ impl<SDK: SovereignAPI> BlendedRuntime<SDK> {
             use fluentbase_sdk::runtime::RuntimeContextWrapper;
             let runtime_context = RuntimeContext::root(params.fuel_limit);
 
-            // For multicall, we need to get the bytecode from the multicall contract
-            // For regular calls, we get the bytecode from the contract context
-            let (_, bytecode) = if is_multicall {
-                let (acc, _) = self.sdk.account(&EXAMPLE_MULTICALL_ADDRESS);
-                let code = self
-                    .sdk
-                    .preimage(&EXAMPLE_MULTICALL_ADDRESS, &acc.code_hash)
-                    .unwrap_or_default();
-                (acc, code)
-            } else {
-                let (acc, _) = self.sdk.account(&contract_context.bytecode_address);
-                let code = self
-                    .sdk
-                    .preimage(&contract_context.bytecode_address, &acc.code_hash)
-                    .unwrap_or_default();
-                (acc, code)
-            };
+            let (account, _) = self.sdk.account(&contract_context.bytecode_address);
+            let bytecode = self
+                .sdk
+                .preimage(&contract_context.bytecode_address, &account.code_hash)
+                .unwrap_or_default();
 
-            // Handle EIP7702 delegation and create preimage adapter
             let preimage_adapter = if let Bytecode::Eip7702(eip7702_bytecode) =
                 Bytecode::new_raw(bytecode)
             {
                 crate::helpers::SdkPreimageAdapter(eip7702_bytecode.delegated_address, &self.sdk)
             } else {
-                crate::helpers::SdkPreimageAdapter(
-                    if is_multicall {
-                        EXAMPLE_MULTICALL_ADDRESS
-                    } else {
-                        contract_context.bytecode_address
-                    },
-                    &self.sdk,
-                )
+                crate::helpers::SdkPreimageAdapter(contract_context.bytecode_address, &self.sdk)
             };
 
-            // Setup runtime context and execute
             let native_sdk = RuntimeContextWrapper::new(runtime_context)
                 .with_preimage_resolver(&preimage_adapter);
             let (fuel_consumed, exit_code) = native_sdk.exec(
@@ -203,27 +175,48 @@ impl<SDK: SovereignAPI> BlendedRuntime<SDK> {
         }
     }
 
+    /// Checks if the function call should be redirected to a native precompiled contract.
+    ///
+    /// When the first 4 bytes of the input (function selector) match a precompile's address prefix,
+    /// returns the corresponding precompile account that should handle the call.
+    ///
+    /// # Arguments
+    /// * `input` - The complete calldata for the function call
+    ///
+    /// # Returns
+    /// * `Some(Account)` - The precompile account if a match is found
+    /// * `None` - If no matching precompile is found or input is too short
+    fn try_resolve_precompile_account(&mut self, input: &[u8]) -> Option<Account> {
+        if input.len() < 4 {
+            return None;
+        };
+
+        if input[..4] == PRECOMPILE_MULTICALL[..4] {
+            let (acc, _) = self.sdk.account(&PRECOMPILE_MULTICALL);
+            return Some(acc);
+        }
+
+        None
+    }
+
     fn exec_bytecode(
         &mut self,
-        context: ContractContext,
+        mut context: ContractContext,
         mut bytecode_account: Account,
         input: Bytes,
         gas: &mut Gas,
         state: u32,
         call_depth: u32,
     ) -> (Bytes, i32) {
-        let mut bytecode = if input.len() >= 4 && input[..4] == MULTICALL_SELECTOR {
-            let (multicall_account, _) = self.sdk.account(&EXAMPLE_MULTICALL_ADDRESS);
-            bytecode_account = multicall_account;
+        if let Some(precompile_account) = self.try_resolve_precompile_account(&input) {
+            bytecode_account = precompile_account;
+            context.bytecode_address = bytecode_account.address;
+        }
 
-            self.sdk
-                .preimage(&bytecode_account.address, &bytecode_account.code_hash)
-                .unwrap_or_default()
-        } else {
-            self.sdk
-                .preimage(&bytecode_account.address, &bytecode_account.code_hash)
-                .unwrap_or_default()
-        };
+        let mut bytecode = self
+            .sdk
+            .preimage(&bytecode_account.address, &bytecode_account.code_hash)
+            .unwrap_or_default();
 
         if let Bytecode::Eip7702(eip7702_bytecode) = Bytecode::new_raw(bytecode.clone()) {
             let (delegated_account, _) = self.sdk.account(&eip7702_bytecode.delegated_address);
