--- examples/multicall/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-edition = "2021"
-name = "fluentbase-example-multicall"
-version = "0.1.0"
-
-[dependencies]
-alloy-primitives = {workspace = true}
-fluentbase-sdk = {workspace = true}
-
-[dev-dependencies]
-alloy-sol-types = {workspace = true, default-features = false}
-hex = "0.4.3"
-
-[lib]
-crate-type = ["cdylib", "staticlib"]
-path = "lib.rs"
-
-[features]
-default = ["std"]
-std = [
-  "fluentbase-sdk/std",
-]
