--- examples/multicall/README.md
@@ -1,3 +1,0 @@
-# Multicall contract
-
-A precompile contract, which allows to call multiple functions in a single transaction.
